requirements.txt
beautifulsoup4==4.13.4
catboost
dnspython
flask==2.3.3
flask-cors==4.0.0
joblib==1.3.2
numpy==2.3.1
pandas==2.3.0
python-dotenv
requests==2.32.4
scikit-learn==1.7.0
tldextract==5.3.0
xgboost==3.0.2